# internal/transport/http/auth/model.gen.go
// Package auth provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package auth

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Error defines model for Error.
type Error struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// OK defines model for OK.
type OK struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// SignInForm defines model for SignInForm.
type SignInForm struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// SignUpForm defines model for SignUpForm.
type SignUpForm struct {
	Email     *openapi_types.Email `json:"email,omitempty"`
	FirstName string               `json:"firstName"`
	LastName  string               `json:"lastName"`
	Password  string               `json:"password"`
	Phone     string               `json:"phone"`
	Username  string               `json:"username"`
}

// UpdateUser defines model for UpdateUser.
type UpdateUser struct {
	Email     openapi_types.Email `json:"email"`
	FirstName string              `json:"firstName"`
	LastName  string              `json:"lastName"`
	Phone     string              `json:"phone"`
	Username  string              `json:"username"`
}

// User defines model for User.
type User struct {
	Email     *openapi_types.Email `json:"email,omitempty"`
	FirstName string               `json:"firstName"`
	LastName  string               `json:"lastName"`
	Phone     string               `json:"phone"`
	Username  string               `json:"username"`
}

// SignInJSONRequestBody defines body for SignIn for application/json ContentType.
type SignInJSONRequestBody = SignInForm

// SignUpJSONRequestBody defines body for SignUp for application/json ContentType.
type SignUpJSONRequestBody = SignUpForm

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = UpdateUser